Add GiamDoc serivce

+++ NhanVien/src/main/java/com/vinorsoft/nhanvien/common/entity/NhanVien.java
@@ -42,7 +42,7 @@ public class NhanVien {
     @Column(length = 10)
     private String chucVuId;
 
-    private double luong;
+    private long luong;
 
     private byte trangThai;
 }
