+++ NhanVien/src/main/java/com/vinorsoft/nhanvien/payload/request/nhanvien/UpdateNhanVienRequest.java
@@ -35,7 +35,7 @@ public class UpdateNhanVienRequest {
     private String chucVuId;
 
     @NotNull(message = "luong is mandatory")
-    private Double luong;
+    private Long luong;
 
     private Byte trangThai;
 }
